+++ source/findLast.js
@@ -4,7 +4,7 @@ export function findLast(predicate, list){
   let index = list.length
 
   while (--index >= 0){
-    if (predicate(list[ index ], index)){
+    if (predicate(list[ index ])){
       return list[ index ]
     }
   }