+++ src/all.js
@@ -2,7 +2,7 @@ export function all(predicate, list){
   if (arguments.length === 1) return _list => all(predicate, _list)
 
   for (let i = 0; i < list.length; i++){
-    if (!predicate(list[ i ], i)) return false
+    if (!predicate(list[ i ])) return false
   }
 
   return true